+++ src/utils/logging/nodeFetch/nodeFetch.ts
@@ -1,3 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
 import { ext } from "../../../extensionVariables";
 import { HttpLogger } from "../HttpLogger";
 import { NodeFetchNormalizer } from "./NodeFetchNormalizer";