+++ test/api/branchDataProvider.test.ts
@@ -1,3 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
 import assert from "assert";
 import { Event, TreeDataProvider, TreeItem } from "vscode";
 import { isWrapper, ResourceModelBase, WorkspaceResource, WorkspaceResourceProvider } from "../../api/src";