+++ test/api/workspaceResourceProvider.test.ts
@@ -1,3 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
 import assert from "assert";
 import { TreeItem } from "vscode";
 import type { ResourceModelBase, WorkspaceResource, WorkspaceResourceProvider } from "../../api/src";