+++ test/wrapFunctionsInTelemetry.test.ts
@@ -1,3 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
 import { IActionContext } from "@microsoft/vscode-azext-utils";
 import assert from "assert";
 import { wrapFunctionsInTelemetry, wrapFunctionsInTelemetrySync } from "../src/utils/wrapFunctionsInTelemetry";